refactor: File structure & magic numbers for users
