+++ src/__tests__/AppModule.test.ts
@@ -2,8 +2,8 @@ import request from 'supertest';
 import { Controller, Get, HttpStatus } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AppModule } from '../application/modules/AppModule';
-import { LoggerMiddleware } from '../common/middleware/LoggerMiddleware';
-import { HttpErrorFilter } from '../common/filters/HttpErrorFilter';
+import { LoggerMiddleware } from '../http_api/middleware/LoggerMiddleware';
+import { HttpErrorFilter } from '../http_api/filters/HttpErrorFilter';
 import { HttpExceptionMessages } from '../common/enums/HttpExceptionMessages';
 import { NewWinstonAdapter } from '../infrastructure/logging/adapters/NewWinstonAdapter';
 import { mockILogger } from './mocks/mockLogAdapter';
