+++ src/application/dtos/user/UserResponseDto.ts
@@ -2,6 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { IsString, MaxLength, MinLength } from 'class-validator';
 import { UserEntity } from '../../../domain/entities/user/UserEntity';
 import { ResponseDto } from '../ResponseDto';
+import { userConstants } from '../../../common/constants/userConstants';
 
 /**
  * This class is responsible for returning a {@link UserEntity} to the client.
@@ -12,8 +13,8 @@ import { ResponseDto } from '../ResponseDto';
 export class UserResponseDto extends ResponseDto {
 	@ApiProperty({ description: 'The unique username of the user entity', uniqueItems: true, required: true })
 	@IsString({ message: 'value must be a string' })
-	@MinLength(3, { message: 'Username must be at least 3 characters long' })
-	@MaxLength(50, { message: 'Username must be at most 50 characters long' })
+	@MinLength(userConstants.minUsernameLength, { message: `Username must be at least ${userConstants.minUsernameLength} characters long` })
+	@MaxLength(userConstants.maxUsernameLength, { message: `Username can be at most ${userConstants.maxUsernameLength} characters long` })
 	username: string;
 
 	constructor(entity: UserEntity) {
