+++ src/common/constants/userConstants.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared constatns for user entities and dto's.
+ */
+export const userConstants = {
+	minUsernameLength: 3,
+	maxUsernameLength: 50,
+};
