+++ src/http_api/controllers/GuardedController.ts
@@ -2,18 +2,18 @@ import { BadRequestException, Body, Param, ParseIntPipe, UseFilters, UseGuards }
 import { Observable } from 'rxjs';
 import { ApiSecurity } from '@nestjs/swagger';
 import { isTruthy } from 'ts-istruthy';
-import { BadRequestExceptionFilter } from '../../common/filters/BadRequestExceptionFilter';
-import { HttpExceptionFilter } from '../../common/filters/HttpExceptionFilter';
-import { NotFoundExceptionFilter } from '../../common/filters/NotFoundExceptionFilter';
-import { NotImplementedExceptionFilter } from '../../common/filters/NotImplementedExceptionFilter';
-import { QueryFailedErrorFilter } from '../../common/filters/QueryFailedErrorFilter';
-import { UnauthorizedExceptionFilter } from '../../common/filters/UnauthorizedExceptionFilter';
-import { PassportJwtAuthGuard } from '../../common/guards/PassportJwtAuthGuard';
-import { AbstractService } from '../services/AbstractService';
-import { CreateDto } from '../dtos/CreateDto';
-import { UpdateDto } from '../dtos/UpdateDto';
-import { ResponseDto } from '../dtos/ResponseDto';
-import { ISseMessage } from '../events/ISseMessage';
+import { BadRequestExceptionFilter } from '../filters/BadRequestExceptionFilter';
+import { HttpExceptionFilter } from '../filters/HttpExceptionFilter';
+import { NotFoundExceptionFilter } from '../filters/NotFoundExceptionFilter';
+import { NotImplementedExceptionFilter } from '../filters/NotImplementedExceptionFilter';
+import { QueryFailedErrorFilter } from '../filters/QueryFailedErrorFilter';
+import { UnauthorizedExceptionFilter } from '../filters/UnauthorizedExceptionFilter';
+import { PassportJwtAuthGuard } from '../guards/PassportJwtAuthGuard';
+import { AbstractService } from '../../application/services/AbstractService';
+import { CreateDto } from '../../application/dtos/CreateDto';
+import { UpdateDto } from '../../application/dtos/UpdateDto';
+import { ResponseDto } from '../../application/dtos/ResponseDto';
+import { ISseMessage } from '../../application/events/ISseMessage';
 import { ILogger } from '../../infrastructure/logging/ILogger';
 import { NewWinstonAdapter } from '../../infrastructure/logging/adapters/NewWinstonAdapter';
