+++ src/http_api/decorators/DeleteEndpoint.ts
@@ -1,6 +1,6 @@
 import { applyDecorators, Delete, HttpCode, HttpStatus } from '@nestjs/common';
 import { ApiOperation, ApiResponse } from '@nestjs/swagger';
-import { HttpExceptionMessages } from '../enums/HttpExceptionMessages';
+import { HttpExceptionMessages } from '../../common/enums/HttpExceptionMessages';
 import { DefaultErrorDecorators } from './DefaultErrorDecorators';
 
 /**
