+++ src/http_api/decorators/SseEndpoint.ts
