+++ src/http_api/filters/NotImplementedExceptionFilter.ts
@@ -1,6 +1,6 @@
 import { Catch, HttpStatus, NotImplementedException } from '@nestjs/common';
 import { AbstractHttpFilter } from './AbstractHttpFilter';
-import { HttpExceptionMessages } from '../enums/HttpExceptionMessages';
+import { HttpExceptionMessages } from '../../common/enums/HttpExceptionMessages';
 import { NewWinstonAdapter } from '../../infrastructure/logging/adapters/NewWinstonAdapter';
 
 @Catch(NotImplementedException)
