+++ src/http_api/filters/UnauthorizedExceptionFilter.ts
@@ -1,6 +1,6 @@
 import { Catch, HttpStatus, UnauthorizedException } from '@nestjs/common';
 import { AbstractHttpFilter } from './AbstractHttpFilter';
-import { HttpExceptionMessages } from '../enums/HttpExceptionMessages';
+import { HttpExceptionMessages } from '../../common/enums/HttpExceptionMessages';
 import { NewWinstonAdapter } from '../../infrastructure/logging/adapters/NewWinstonAdapter';
 
 @Catch(UnauthorizedException)
