+++ src/http_api/guards/PassportJwtAuthGuard.ts
